Added HP and Damage to the characters and minions.
Damage works as it should, but still testing. When you attack the sphere
in the corner enough times, it will disappear. (JUST FOR TESTING)

A known bug is that the  proper player choice is still not correct at
times. so when the necromancer summons the wraith, and you click move,
it still thinks you're on the necromancer. The necromancer's hitbox is a
little strange